pyproject.toml
[project]
name = "ai-of-might-and-delusion"
version = "0.1.0"
description = "HOMM3 (vcmi) map generator using LLMs"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "click>=8.1.8",
    "docker>=7.1.0",
    "easyprocess>=1.1",
    "gitpython>=3.1.45",
    "hydra-core>=1.3.2",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "pyautogui>=0.9.54",
    "pydantic-ai-slim[anthropic,openai]>=0.8.1",
    "python-dotenv>=1.1.1",
    "pyvirtualdisplay>=3.0",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
